third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-mouse-events-between-frames-expected.txt

Verifies that mouse events are fired on tapping of gesture across frames in the correct order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

--- Tap tests ---
-- sending gestureTap to green --
PASS fired event is (yellow, mouseover).
PASS fired event is (yellow, mouseenter).
PASS fired event is (greenyellow, mouseover).
PASS fired event is (greenyellow, mouseenter).
PASS fired event is (limegreen, mouseover).
PASS fired event is (limegreen, mouseenter).
PASS fired event is (green, mouseover).
PASS fired event is (green, mouseenter).
-- sending gestureTap to green --
There should not be any mouse event.
-- sending gestureTap to hotpink --
PASS fired event is (green, mouseout).
PASS fired event is (green, mouseleave).
PASS fired event is (limegreen, mouseout).
PASS fired event is (limegreen, mouseleave).
PASS fired event is (greenyellow, mouseout).
PASS fired event is (greenyellow, mouseover).
PASS fired event is (pink, mouseover).
PASS fired event is (pink, mouseenter).
PASS fired event is (hotpink, mouseover).
PASS fired event is (hotpink, mouseenter).
-- sending gestureTap to hotpink --
There should not be any mouse event.
-- sending gestureTap to green --
PASS fired event is (hotpink, mouseout).
PASS fired event is (hotpink, mouseleave).
PASS fired event is (pink, mouseout).
PASS fired event is (pink, mouseleave).
PASS fired event is (greenyellow, mouseout).
PASS fired event is (greenyellow, mouseover).
PASS fired event is (limegreen, mouseover).
PASS fired event is (limegreen, mouseenter).
PASS fired event is (green, mouseover).
PASS fired event is (green, mouseenter).
PASS successfullyParsed is true

TEST COMPLETE
